Add describe()  method
