+++ package.json
@@ -27,7 +27,7 @@
     "webpack-dev-server": "^4.2.1"
   },
   "scripts": {
-    "test": "jest",
+    "test": "jest --coverage",
     "start": "webpack serve --open",
     "build": "webpack"
   },
